requirements.txt
Django==1.8.5
Pillow==2.9.0
PyJWT==1.4.0
PyYAML==3.12
Whoosh==2.7.4
app-version==0.2.1
backports.ssl-match-hostname==3.5.0.1
blessed==1.9.5
botocore==1.4.53
cement==2.8.2
click==6.2
colorama==0.3.7
decorator==4.0.6
dj-stripe==0.8.0
django-activity-stream==0.6.0
django-allauth==0.23.0
django-appconf==1.0.1
django-braces==1.8.1
django-classy-tags==0.6.2
django-contrib-comments==1.6.1
django-envelope==1.1
django-geojson==2.8.1
django-haystack==2.4.1
django-localflavor==1.2
django-model-utils==2.4
django-permission==0.9.2
djangorestframework==3.3.2
djangorestframework-gis==0.10.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
funcsigs==0.4
geocoder==1.8.0
geopy==1.11.0
jmespath==0.9.0
jsonfield==1.0.3
mock==1.3.0
oauthlib==1.0.3
pathspec==0.3.4
pbr==1.8.1
psycopg2==2.6.1
python-dateutil==2.5.3
python-openid==2.2.5
pytz==2015.7
ratelim==0.1.6
requests==2.7.0
requests-oauthlib==0.5.0
semantic-version==2.5.0
six==1.9.0
sorl-thumbnail==12.3
stripe==1.27.1
texttable==0.8.4
tolerance==0.1.2
urllib3==1.14
wcwidth==0.1.7
websocket-client==0.37.0
wsgiref==0.1.2
